+++ include/wisey/IRGenerationContext.hpp
@@ -81,7 +81,7 @@ namespace wisey {
     /**
      * Run compiled IR code and return the result
      */
-    int runCode(int argc, char** argv);
+    long runCode(int argc, char** argv);
     
     /**
      * Return the main module
